requirements.txt
langchain>=0.2.6
langchain-openai>=0.1.14
langchain-core>=0.2.16
langchain-community>=0.2.0
langchain-text-splitters>=0.3.0
pydantic>=2.7.0
python-dotenv>=1.0.1
typer>=0.12.5
rich>=13.7.1
streamlit>=1.41.1
faiss-cpu>=1.8.0
unstructured>=0.14.6
PyPDF2>=3.0.1
openai
